refactor (front): criando o useClients
